Move "Settings" out from "Konfig" to separate class
The "Settings" class has been moved out from the "Konfig" class into it's own package, and renamed to "KonfigSettings". The "UnknownEntryBehaviour" enum that previously resided inside of the "Settings" nested class has also been moved out from it.

The entire structure of the settings class has also been changed, as before it was designed to be "mutable", which would could allow for some troublesome scenarios, so it has been changed to be immutable.

Signed-off-by: Olivki <oliver@berg.moe>