Add more ways to add layers and entries
Entries:
A "addEntries(...)" function has been added, which allows the user to
add a collection of entries to the specified layer.

Layers:
A function for adding an already established collection of layers to
a specified layer has been added, and functions for creating and then
adding the created layers from just a string have been added.

Signed-off-by: Olivki <oliver@berg.moe>